LPS-117839 Move implementations to service module
