+++ modules/apps/translation/translation-service/build.gradle
@@ -4,17 +4,30 @@ buildService {
 }
 
 dependencies {
+	compileInclude group: "com.fasterxml.woodstox", name: "woodstox-core", version: "6.2.1"
+	compileInclude group: "com.ibm.icu", name: "icu4j", version: "65.1"
+	compileInclude group: "net.sf.okapi", name: "okapi-core", version: "1.39.0"
+	compileInclude group: "net.sf.okapi.filters", name: "okapi-filter-autoxliff", version: "1.39.0"
+	compileInclude group: "net.sf.okapi.filters", name: "okapi-filter-xliff", version: "1.39.0"
+	compileInclude group: "net.sf.okapi.filters", name: "okapi-filter-xliff2", version: "1.39.0"
+	compileInclude group: "net.sf.okapi.lib", name: "okapi-lib-xliff2", version: "1.39.0"
+	compileInclude group: "org.codehaus.woodstox", name: "stax2-api", version: "4.2.1"
+
 	compileOnly group: "com.liferay", name: "biz.aQute.bnd.annotation", version: "4.2.0.LIFERAY-PATCHED-1"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "javax.portlet", name: "portlet-api", version: "3.0.1"
 	compileOnly group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
+	compileOnly project(":apps:dynamic-data-mapping:dynamic-data-mapping-api")
 	compileOnly project(":apps:info:info-api")
+	compileOnly project(":apps:journal:journal-api")
 	compileOnly project(":apps:portal:portal-aop-api")
 	compileOnly project(":apps:portal:portal-instance-lifecycle-api")
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	compileOnly project(":apps:translation:translation-api")
+	compileOnly project(":core:osgi-service-tracker-collections")
+	compileOnly project(":core:petra:petra-function")
 	compileOnly project(":core:petra:petra-io")
 	compileOnly project(":core:petra:petra-lang")
 	compileOnly project(":core:petra:petra-sql-dsl-api")
