+++ modules/apps/translation/translation-service/src/main/java/com/liferay/translation/internal/exporter/XLIFF20InfoFormTranslationExporter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.translation.web.internal.exporter;
+package com.liferay.translation.internal.exporter;
 
 import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldValue;
@@ -40,10 +40,10 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alejandro Tardín
  */
 @Component(
-	property = "content.type=application/x-xliff+xml",
+	property = "content.type=application/xliff+xml",
 	service = TranslationInfoItemFieldValuesExporter.class
 )
-public class XLIFF12InfoFormTranslationExporter
+public class XLIFF20InfoFormTranslationExporter
 	implements TranslationInfoItemFieldValuesExporter {
 
 	@Override
@@ -55,29 +55,23 @@ public class XLIFF12InfoFormTranslationExporter
 		Document document = SAXReaderUtil.createDocument();
 
 		Element xliffElement = document.addElement(
-			"xliff", "urn:oasis:names:tc:xliff:document:1.2");
+			"xliff", "urn:oasis:names:tc:xliff:document:2.0");
 
-		xliffElement.addAttribute("version", "1.2");
+		xliffElement.addAttribute(
+			"srcLang", LocaleUtil.toBCP47LanguageId(sourceLocale));
+		xliffElement.addAttribute(
+			"trgLang", LocaleUtil.toBCP47LanguageId(targetLocale));
+		xliffElement.addAttribute("version", "2.0");
 
 		Element fileElement = xliffElement.addElement("file");
-
-		fileElement.addAttribute("datatype", "plaintext");
 
 		InfoItemReference infoItemReference =
 			infoItemFieldValues.getInfoItemReference();
 
 		fileElement.addAttribute(
-			"original",
+			"id",
 			infoItemReference.getClassName() + StringPool.COLON +
 				infoItemReference.getClassPK());
-
-		fileElement.addAttribute(
-			"source-language", LocaleUtil.toBCP47LanguageId(sourceLocale));
-		fileElement.addAttribute(
-			"target-language", LocaleUtil.toBCP47LanguageId(targetLocale));
-		fileElement.addAttribute("tool", "Liferay");
-
-		Element bodyElement = fileElement.addElement("body");
 
 		Collection<InfoFieldValue<Object>> infoFieldValues =
 			infoItemFieldValues.getInfoFieldValues();
@@ -89,21 +83,19 @@ public class XLIFF12InfoFormTranslationExporter
 				continue;
 			}
 
-			Element transUnitElement = bodyElement.addElement("trans-unit");
+			Element unitElement = fileElement.addElement("unit");
 
-			transUnitElement.addAttribute("id", infoField.getName());
+			unitElement.addAttribute("id", infoField.getName());
 
-			Element sourceElement = transUnitElement.addElement("source");
+			Element segmentElement = unitElement.addElement("segment");
 
-			sourceElement.addAttribute(
-				"xml:lang", fileElement.attributeValue("source-language"));
+			Element sourceElement = segmentElement.addElement("source");
+
 			sourceElement.addCDATA(
 				_getStringValue(infoFieldValue.getValue(sourceLocale)));
 
-			Element targetElement = transUnitElement.addElement("target");
+			Element targetElement = segmentElement.addElement("target");
 
-			targetElement.addAttribute(
-				"xml:lang", fileElement.attributeValue("target-language"));
 			targetElement.addCDATA(
 				_getStringValue(infoFieldValue.getValue(targetLocale)));
 		}
@@ -115,7 +107,7 @@ public class XLIFF12InfoFormTranslationExporter
 
 	@Override
 	public String getMimeType() {
-		return "application/x-xliff+xml";
+		return "application/xliff+xml";
 	}
 
 	private String _getStringValue(Object value) {
