+++ modules/apps/translation/translation-service/src/main/java/com/liferay/translation/internal/importer/XLIFFInfoFormTranslationImporter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.translation.web.internal.importer;
+package com.liferay.translation.internal.importer;
 
 import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldValue;
